Created common base class for testing different compilers
This commit serves the purposes of clearly isolating the initial refactoring
needed to run the same tests against both the vanilla Scala compiler and the
Triplequote Hydra compiler.

The next commit introduces the changes required to have such flexibility in the
test suite.
